contract_renewal/backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf2
requests
python-dotenv 
firebase-admin 
azure-storage-blob